Create lists of ValidationError only when necessary
This commit reduces memory usage by only creating a list of
ValidationException objects when the first one in encountered. JSONs
that are valid will gain the most in terms of memory usage, as no lists
will be created. JSONs that are only partially valid will still see
memory usage improvements.