Fixing regex handling in `"patternProperties"`
Changing `ObjectSchema` to represent `patternProperties` as a `Map<Regexp, Schema>` instead of `Map<Pattern, Schema>`
This is a follow-up change of introducing `Regexp` and `RegexpFactory` interfaces in `1.9.0`.

Keeping but deprecating existing accessor & builder methods of `ObjectSchema` which consume / return a `Pattern`, converting from/to `Regexp`.

In the end, changing `ObjectSchemaLoader` to instantiate a `Regexp` instance using the configured `RegexpFactory` instead of attempting to use RE2J directly.

Also bumping version number to 1.9.1, since it is going to be a quick bugfix release ASAP.