initial steps to better track original locations of remote-loaded schemas
So far it only compiles but hundreds of tests fail due to obvious reasons.

A new `SchemaLocation` class is introduced, extending the public interface of the library. The `SchemaLocation` tracks both
 * the source URL of the root document of the remote json document containing the schema (if present),
 * and the json pointer denoting the schema within the root document.

Schema instances now track the location as a `SchemaLocation` instead of the previous representation (only a json pointer
as a `String`).

Also changing the representation of the json pointer from `List<String>` to `SchemaLocation` at the following places in
the `loader` package:
 * `LoadingState#pointerToCurrentObj`
 * `SchemaLoader#pointerToCurrentObj`

Also minor code cleanups in `ValidationException`.