making it possible to pre-define a schema by URI before schema loading
Reference: section 8.3.1 of json schema core draft-7 says:
"Such URIs and schemas can be supplied to an implementation prior to processing instances"

So here if the caller associates a raw schema to an URI using `SchemaLoaderBuilder#registerSchemaByURI()` then the loader
will recognize this URI and load the associated raw schema without hitting the network or any external protocols.

This was originally requested in #38.