omitting stacktraces for internal exceptions
The purpose of this change is to avoid stacktrace generation for most `ValidationException` instances, since it can
come with significant performance improvements in some usecases, see #318 .

An `InternalValidationException` is introduced which is a subclass of `ValidationException` but it omits the stacktrace
generation by overriding `fillInStackTrace()` to be no-op. This is now thrown everywhere internally. At the end of the
validation process the `DefaultValidator` catches the possibly thrown `InternalValidationException` and copies it into a
`ValidationInstance` and throws it, so from the caller point of view an exception with a proper stacktrace is thrown. This
"public" `ValidationException` is necessary to avoid any potential BC breaks.

Tests updated. A good number of unittests and all of the integration tests verify that only root exceptions (and no causing
exceptions) have non-empty stacktraces. There is some code duplication between the unit and integ test codebase, this is
to be cleaned up a bit in the future (needs a separate testutil module which can be depended on by both `core` and
`tests/vanilla`).