fixing #327
Fixing issue #327 (determining id/$id depending on schema spec version). The $schema value is preserved after parsing as an
unprocessedProperty so we use that. The deduced spec version is stored in a visitor field, so the first found $schema value
is considered as the spec version for the entire schema document.

Also:
 * printing common properties for EnumSchema
 * changing ToStringVisitor to use its own #printSchemaMap() because the previously used JSONPrinter#printSchemaMap() created